Add the storage variables to faucet contract

+++ contracts/Faucet.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.4.22 <0.9.0;
 
-contract Faucet{
+contract Faucet {
     // Storage variables
-    uint public funds = 1000;
-    int public counter  = -10;
+    uint256 public funds = 1000;
+    int256 public counter = -10;
 }
